Integer-size warning related fix
git-svn-id: http://svn.code.sf.net/p/cmusphinx/code/trunk@12659 94700074-3cef-4d97-a70e-9c8c206c02f5

+++ sphinx3/src/libs3decoder/libAPI/utt.c
@@ -82,7 +82,7 @@ utt_decode(void *data, utt_res_t * ur, int32 sf, int32 ef, char *uttid)
     /* Convert input file to cepstra if waveform input is selected */
     if (cmd_ln_boolean_r(config, "-adcin")) {
         int16 *adcdata;
-        int32 nsamps = 0;
+        size_t nsamps = 0;
 
         if ((adcdata = bio_read_wavfile(cmd_ln_str_r(config, "-cepdir"),
     				        ur->uttfile,
